Refactor renderer layout and add renderpasses
Move renderer subfolders into top-level src (descriptors, pipelines, shader) and update project files/filters and include paths accordingly. Add renderer_renderpasses.cpp with draw-image and swapchain renderpass creation, and wire new VkRenderPass members and creation methods into Renderer. Implement image transition helpers and overloads in vk_helper, expose image/format accessors and image count in VulkanBackend/vk_swapchain, and add renderpass tracking in vk_util. Also adjust FRAME_OVERLAP checks, change semaphore loop index type, and update offscreen target usage and deletion queue handling.

+++ Blasphemy Engine/src/vulkan_backend/vk_swapchain.cpp
@@ -78,7 +78,7 @@ bool VulkanSwapchain::create(VkPhysicalDevice physicalDevice, VkDevice device, V
 
 	renderSemaphores.resize(imageCount);
 	VkSemaphoreCreateInfo semaphoreCreateInfo = vkhelper::semaphore_create_info();
-	for (uint32_t i = 0; i < imageCount; i++) {
+	for (size_t i = 0; i < imageCount; i++) {
 		VK_CHECK(vkCreateSemaphore(device, &semaphoreCreateInfo, nullptr, &renderSemaphores[i]));
 	}
